# Avoid deepcopy in `_tell_with_warning`