## Motivation
Improve the performance of `_tell_with_warning` by reducing unnecessary deep copies.

## Description of the Changes
- Replaced `deepcopy` with a shallow copy in `_tell_with_warning` to reduce overhead when returning a `FrozenTrial`.
- A `deepcopy` is performed only in contexts where the `FrozenTrial` might be modified, such as when it is passed into callbacks or used in `study.tell()`.