@sawa3030 2025-05-09
## Performance Improvement
The test was conducted by alternating between the `master` branch and this PR for 10 iterations.  The results show a slight improvement in the cumulative time (`cumtime`) of `_run_trial`, as expected.
note: Because the performance difference is subtle, results from a single iteration was heavily influenced by external factors and are therefore not reliable on their own.

| Branch    | Average Cumtime |
|-----------|------------------|
| master    | 16.6892 ± 0.43999      |
| this PR   | 16.4755 ± 0.48987     | 

The code used for testing is below.
```
import optuna
import cProfile

def objective(trial: optuna.Trial) -> float:
    x = trial.suggest_float("x", -1, 1)
    y = trial.suggest_int("y", -1, 1)
    return x**2 + y

study = optuna.create_study()
profiller = cProfile.Profile()
profiller.run('study.optimize(objective, n_trials=500)')
```