@not522 2025-05-12
There is a significant speed difference with `RandomSampler`.

- master `1.50s user 0.06s system 93% cpu 1.681 total`
- PR `1.34s user 0.07s system 87% cpu 1.622 total`

```python
import optuna

def objective(trial: optuna.Trial) -> float:
    x = trial.suggest_float("x", -1, 1)
    y = trial.suggest_int("y", -1, 1)
    return x**2 + y

study = optuna.create_study(sampler=optuna.samplers.RandomSampler())
study.optimize(objective, n_trials=10000)
```