@sawa3030 2025-05-14
@not522 Thank you for the suggestion. I have opened a new PR [#6082](https://github.com/optuna/optuna/pull/6082), so I will close this PR for now. Once #6082 is merged, I will reopen this PR accordingly.