@sawa3030 2025-05-23
Let me reopen this PR since [#6082](https://github.com/optuna/optuna/pull/6082) is merged into master